Add -https flag to the web server to disable TLS
This flag is meant to be used when running behind an HTTPS proxy such as
jrcs/letsencrypt-nginx-proxy-companion. Although such proxies can be configured
to connect via HTTPS on the backend, the use of autocert (layered LetsEncrypt)
is not meaningful.

Running completely unencrypted is not supported (for example, generated URLs
are all HTTPS).