docker-compose: Duplicate config across configurations, fix running
This effectively fixes the breakage of the proxied configuration introduced by
3ab0453ac8f. (If you are running a reverse proxy, Docker will fail to start
YAGPDB as the host ports will already be in use.)

The cascading configuration was supposed to reduce the amount of maintenance,
but does so at the cost of increased complexity. Some items, such as the port
mappings, can't be overridden, only merged, so we either have to _always_
require a hierarchy of configs (as-is currently the case) or we flatten the
hierarchy and occasionally have to make changes in multiple files. As the
former has already resulted in one breakage, we may as well go with simplicity.

(An alternative case: a little bit of pain on the rare config change or a
little bit of pain every time someone starts the server. This change reduces
overall pain.)

Other changes introduced in merging the config:
* Change proxied to also use the caskd image
* Fix dev ports to work like the current non-proxied config
* Migrate command to non-abbreviated for improved diff output