docker-compose: Duplicate config across configurations, fix running (#474)
* Docker: Pin Postgres major version

The only method to migrate data between major Postgres vesrions is a backup and
restore. This provides no clean path for users to rebuild their server so let's
pin the version to reduce surprises.

Although version 10 was current when this config was originally checked in,
version 11 has been the latest for 9 months and I am assuming more users are
using that "latest".

With version 12 coming out soon, it seems like a good time to resolve this
problem

For users discovering this commit message that wish to upgrade, the following
gist can be useful:
https://gist.github.com/LoganK/7ceb1506102f5d469556c69620a3d1c2

(Remember to change your volume name in your docker-compose file.)

* docker-compose.dev: Add the host port mapping config

This brings it in line with docker-compose.yml and allows running as a live server.

* docker-compose.dev: Migrate to Unix EOL for consistency

* docker-compose: Duplicate config across configurations, fix running

This effectively fixes the breakage of the proxied configuration introduced by
3ab0453ac8f. (If you are running a reverse proxy, Docker will fail to start
YAGPDB as the host ports will already be in use.)

The cascading configuration was supposed to reduce the amount of maintenance,
but does so at the cost of increased complexity. Some items, such as the port
mappings, can't be overridden, only merged, so we either have to _always_
require a hierarchy of configs (as-is currently the case) or we flatten the
hierarchy and occasionally have to make changes in multiple files. As the
former has already resulted in one breakage, we may as well go with simplicity.

(An alternative case: a little bit of pain on the rare config change or a
little bit of pain every time someone starts the server. This change reduces
overall pain.)

Other changes introduced in merging the config:
* Change proxied to also use the caskd image
* Fix dev ports to work like the current non-proxied config
* Migrate command to non-abbreviated for improved diff output