Modify timezone setting procedure to behave more properly (see #731)
Formerly, if the user specified a zone name that's a subset of another
one, no action would occur. For example, "setz America/Dawson" would
inform the user that zones called "America/Dawson" and
"America/Dawson_Creek" were available, but would not set the desired
zone. See #731 in jonas747/yagpdb for details.

Now, if multiple matching zones are available but one is an exact match,
it will be set, while also informing the user that other matching zones
are available.