Fix unwarranted assumption, use a set to store matched timezones
Previous code relied on the assumption that, if an exact match occurred, it would be at the beginning of the list of matched timezones. This commit fixes that by switching to a set for matched zone storage (which disallows duplicate values) and testing for membership, rather than simply checking the first item as before.