streaming: ignore other bots
Fairly trivial, bots shouldn't get the role as this feature
is technically meant only for members (see announcement message).
This pull request extends the check for the presence by a check whether the
member in question is not a bot - if so, continue and grant the role.

Let me know when I missed something.