Docker Improvements again (#993)
* Update Dockerfile

Revert the move of copy operations to builder from the last Dockerfile PR, and instead run them in the final image again to make use of image layer download caching. 

Having separate image layers for static/ and templates/ in the final image allows us to cache these when pulling new images from Dockerhub. This means we only need to re-download the image layer with the binary and don't need to re-download the image layers for static/ and templates/ 
This reduces the total download traffic when pulling new images, due to static being close to the size of the binary itself.
However, we leave the copy-templates operation in the builder so we only have one image layer for all templates.

Removed yagpdb/cmd/yagpdb/posts due to it not being needed anymore.

Reduced the wall of text for the RUN apk comment to acceptable levels.

* Update app.example.env

Comment out Twitter credentials, due to them causing the bot to crash if not edited and run with the default placeholder keys instead.

* Update Dockerfile

Changed the "COPY go.mod go.sum" line from "." to "./" since docker build requires it while docker buildx can run without it.
"When using COPY with more than one source file, the destination must be a directory and end with a /"