cc: double-count newlines
Previously, there were cases where the character counter displayed a value
below the limit of 10,000, but errored upon saving. Fix this discrepancy
by double-counting newlines.

The root cause of the issue is that the value of the response textarea on
the frontend contains "\n", while the data received on the backend contains "\r\n",
for reasons unknown to me.

Thanks to Mr. T for helping me figure this out :-)