Several improvements to CCs  (#1104)
* amod: cache invites

* simplify range loop

* use const instead of hardcoded 60 on inviteCacheGC

* use int64 instead of time.time to use less memory

* requested changes

* add templates to pin and unpin messages

* adding limits and merging both template funcs

* fixes and requested changes

* add getPinCount template

* requested fixes

* templates: cube root function

tmplCbrt as context function

* Update general.go

cbrt function

* Update cp_logs.html

Changed the name of the current developer

* templates: add [S]Dict.HasKey method

Add Dict.HasKey and SDict.HasKey, for checking whether the dictionary contains an entry for a particular key.

Presently, the status quo for checking whether a dictionary contains some key is to look up the entry and see if the value is truthy:

{{ if $d.Get $k }}
    {{/* $d has an entry for $k */}}
{{ end }}

The HasKey method improves on this in two ways:

1) It makes the code more self-documenting -- $d.HasKey $k is clear in intent.
2) It handles edge cases better -- if the dictionary can contain entries with zero values (e.g. 0, false) then "if $d.Get $k" may erroneously report that $d has no entry for $k.

* templates: prevent construction of cyclic data

* cc: use monospace font for response box

Co-authored-by: Luca Z. <l-zeuch@email.de>

* cc: double-count newlines

Previously, there were cases where the character counter displayed a value
below the limit of 10,000, but errored upon saving. Fix this discrepancy
by double-counting newlines.

The root cause of the issue is that the value of the response textarea on
the frontend contains "\n", while the data received on the backend contains "\r\n",
for reasons unknown to me.

Thanks to Mr. T for helping me figure this out :-)

* templates: add snowflakeToTime, loadLocation funcs

* templates: wrap SnowflakeToTime to use UTC

...instead of local time, which SnowflakeToTime returns.

* templates: accept interface{} for snowflakeToTime

* templates: add weekNumber

Co-authored-by: Pedro Pessoa <pedro_santos_40@hotmail.com>
Co-authored-by: LemmeCry <67655446+LemmeCry@users.noreply.github.com>
Co-authored-by: caubert <caubert@gmail.com>
Co-authored-by: Black Wolf <TheHDCrafterGaming@gmail.com>
Co-authored-by: Joe <56809242+Jo3-L@users.noreply.github.com>
Co-authored-by: Luca Z. <l-zeuch@email.de>