templates: support threads in `"channel"` carg
This was brought up on the support server, where a user tried to give a
thread as channel argument using `parseArgs` in their custom command.

Arguably, the current behaviour can be considered intended, however I
personally believe that adding another (if not two more) variations like
`"thread"` and `"channel-or-thread"` are undesirable and add more
complication than necessary.

The normal user would expect `parseArgs` to accept also threads as a
`"channel"` carg, and this commit implements exactly that behaviour.