General Bug Fixes and Improvements (#1164)
* amod: cache invites

* simplify range loop

* use const instead of hardcoded 60 on inviteCacheGC

* use int64 instead of time.time to use less memory

* requested changes

* add urlunescape func

* add templates to pin and unpin messages

* adding limits and merging both template funcs

* fixes and requested changes

* add getPinCount template

* requested fixes

* templates: cube root function

tmplCbrt as context function

* Update general.go

cbrt function

* Update cp_logs.html

Changed the name of the current developer

* templates: add [S]Dict.HasKey method

Add Dict.HasKey and SDict.HasKey, for checking whether the dictionary contains an entry for a particular key.

Presently, the status quo for checking whether a dictionary contains some key is to look up the entry and see if the value is truthy:

{{ if $d.Get $k }}
    {{/* $d has an entry for $k */}}
{{ end }}

The HasKey method improves on this in two ways:

1) It makes the code more self-documenting -- $d.HasKey $k is clear in intent.
2) It handles edge cases better -- if the dictionary can contain entries with zero values (e.g. 0, false) then "if $d.Get $k" may erroneously report that $d has no entry for $k.

* templates: prevent construction of cyclic data

* cc: use monospace font for response box

Co-authored-by: Luca Z. <l-zeuch@email.de>

* cc: double-count newlines

Previously, there were cases where the character counter displayed a value
below the limit of 10,000, but errored upon saving. Fix this discrepancy
by double-counting newlines.

The root cause of the issue is that the value of the response textarea on
the frontend contains "\n", while the data received on the backend contains "\r\n",
for reasons unknown to me.

Thanks to Mr. T for helping me figure this out :-)

* templates: add snowflakeToTime, loadLocation funcs

* templates: wrap SnowflakeToTime to use UTC

...instead of local time, which SnowflakeToTime returns.

* templates: accept interface{} for snowflakeToTime

* templates: add weekNumber

* templates: implement json.Marshaler for Dict

* added "lily from at&t" to things to throw

* Disable autorole duration input field if the OnlyOnJoin flag is enabled

* templates: support threads in `"channel"` carg

This was brought up on the support server, where a user tried to give a
thread as channel argument using `parseArgs` in their custom command.

Arguably, the current behaviour can be considered intended, however I
personally believe that adding another (if not two more) variations like
`"thread"` and `"channel-or-thread"` are undesirable and add more
complication than necessary.

The normal user would expect `parseArgs` to accept also threads as a
`"channel"` carg, and this commit implements exactly that behaviour.

* Update TimeTriggerInterval to 5 in interval test cases

* Add test cases for the TimeTriggerInterval validation

* Added "couples therapy" to throwthings.go

Added "couples therapy" to list of things to throw for standard command "throw"

* Add API to handle update and run now for interval custom commands

* bumped dcmd

Co-authored-by: Pedro Pessoa <pedro_santos_40@hotmail.com>
Co-authored-by: LemmeCry <67655446+LemmeCry@users.noreply.github.com>
Co-authored-by: caubert <caubert@gmail.com>
Co-authored-by: Black Wolf <TheHDCrafterGaming@gmail.com>
Co-authored-by: Joe <56809242+Jo3-L@users.noreply.github.com>
Co-authored-by: Luca Z. <l-zeuch@email.de>
Co-authored-by: savage4618 <savage4618@users.noreply.github.com>
Co-authored-by: Ankur P <98815360+ankurp-bst@users.noreply.github.com>
Co-authored-by: Ashish Jhanwar <ashishjh-bst@users.noreply.github.com>