common/util: remove optional protocol from link regex
The current implementation of link validation is, although technically
correct, too sensitive; e.g. it considers an ellipsis (`...`) a valid
link.

This commit modifies the link validator regexp to only consider URLs
with a leading protocol. This should cut down on the reported
false-positives.

The change is justified by the fact that URLs with no leading protocol
aren't "clickable" in Discord either, and as such having it optional is
an anti-pattern, despite the technically correct implemementation.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>