customcommands: handle responses over 2000 chars
This commit updates the `customcommands` command to handle responses
over 2000 characters in the way to instead send them to a file.

This was a suggestion on the support server, the relevant link follows:
https://discord.com/channels/166207328570441728/356486960417734666/927246757501480990

I've tested my changes for single responses over and under 2000
characters, as well as multiple responses, over and under 2000 together
respectively.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>