customcommands: add `raw` switch
Commit 3054fe9 (customcommands: handle responses over 2000 chars) via
PR #1294 introduced a regression with its automagic handling of responses
over 2000 characters, breaking custom commands reliant on the original
behaviour of dumping the code.

This commit attempts to fix that regression by adding a `-raw` switch so
that users relying on the original behaviour can force that for their
custom command """backend""", whilst we still keep a nice user-facing
experience.

Closes #1312

Signed-off-by: Luca Zeuch <l-zeuch@email.de>