refactor(moderation): Remove if from default DM
The Template Script previosuly featured the check, `if .Reason`.
But, `.Reason` returns the string, `"(No reason specified)"` if a reason
wasn't provided when the moderation action was triggered, making this
check redundant.

Additionally, the declaration of the constant used a multi-variable
declaration style, for no dicernable reason, so this has been changed to
a single variable declaration, and with the simplification of the
Template Script, this commit also switches to using a single-line
string, using the newline character.