frontend: add server name to tab title
Add the server name to the tab title if a server's control panel is
open.

Though this is a rather small change, it is useful for those who have
multiple tabs open with different servers -- that way, one can easily
differentiate between open tabs.

I'm not sure how many actually do this, but it was bugging me enough to
whip up a patch, so that must mean something. :)

Signed-off-by: Luca Zeuch <l-zeuch@email.de>