common/templates: add "indent" flag to `tmplJson`
Add an "indent" flag to `tmplJson` in form of a variadic argument.

Generally, users use the `json` function to look at JSON -- with proper
indentation, that becomes like 100 times easier to do.

As to not break backwards compatibility, this flag in form of a boolean
(true/false) is optional and will (obviously) default to false.

The switch statement is basically a copy from the `KindOf` function in
the same file, so there shouldn't be any style issues with it.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>