lib/dcmd: don't accept any number when Mix == Max (#1462)
Stop accepting any number when the specified Min and Max are equal to
each other. Instead, return an OutOfRangeError.

Apply this to all number args we have (i.e. IntArg and FloatArg).

Expected breakage should be low to none. One would generally expect that
when specifying a range with the same Min as Max that only that number
is to be accepted.

See also:
https://canary.discord.com/channels/166207328570441728/356486960417734666/1010540772367540284

Signed-off-by: Luca Zeuch <l-zeuch@email.de>