feat(youtube): Redesign parseYtUrl
This commit redesigns `Plugin.parseYtUrl`, making use of the
`net/url.Url` struct to extract out key elements of YouTube URLs,
minimising the utilisation of RegEx in the method, alongside simplifying
remaining expressions.

Additionally, this adds the `Plugin.parseYtVideoID`
method, to extract some of the parsing logic out of `Plugin.parseYtUrl`,
and introduces support for URLs to YouTube shorts.