customcommands: init work on API
* Add API routes
  For now, we just use web.RequireActiveServer as middleware,
  though that definitely needs an API-friendlier way.

* Implement all GET endpoints
  * `/api/:serverid/customcommands`
  * `/api/:serverid/customcommands/:id`

* Implement premium check middleware
  This works pretty well.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

customcommands: remove "valid" tags from struct

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

customcommands: rename `ok` to `isPremium`

Rename `ok` to `isPremium` to make the code a little clearer on what is
going on.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

customcommands: remove APICustomCommand struct

Internal discussion concluded that this is a very stupid idea and should
not be attempted.

Instead, both accept and send the full custom command struct that is
already in use.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>