feat(youtube): Refactor YouTube matching logic (#1466)
* feat(youtube): Redesign parseYtUrl

This commit redesigns `Plugin.parseYtUrl`, making use of the
`net/url.Url` struct to extract out key elements of YouTube URLs,
minimising the utilisation of RegEx in the method, alongside simplifying
remaining expressions.

Additionally, this adds the `Plugin.parseYtVideoID`
method, to extract some of the parsing logic out of `Plugin.parseYtUrl`,
and introduces support for URLs to YouTube shorts.

* feat(youtube): Update getYtChannel signiature

Updates the signature of `Plugin.getYtChannel`, to work the the
redesigned `Plugin.parseYtUrl` method.

* feat(youtube): Update HandleNew with url.Url

Updates the `Plugin.HandleNew` to use `net/url.Url`, inline with the new
signature of `Plugin.getYtChannel`. This includes updating some variable
names, to avoid blocking the use of the `net/url` package.

* refact(youtube): Add more informative error msgs

Updates `Plugin.parseYtUrl`, using `fmt.Errorf` to include the URL
element, which resulted in it failing to be parsed.