feat(youtube): Add videoID newtype
`videoID` is a newtype, that encapsulates strings representing
YouTube video ID, and features an implementation of the
`getChannelList` method, which fetching data about the video
it refers to, using it to add the channel ID to a `youtube.ChannelsListCall`.

This commit also adds the `listParts` var, as this will be a common
value, also used by other fetching methods.