refact(youtube): Swap \"%s\" for %q
In commit `859d1e6`, I wrapped `%s` flags in error formatting messages,
in `""`, to ensure empty strings would still be visible in error
messages.

After some discussion though, it was decided that using the `%q` flag
instead, was preferable, as this would achieve a similar effect, but
would ensure any characters which may result in confusing formatting
when printing the error, such as newlines and tabs, would instead
be replaced by their escape sequence equivilents, such as `\n` and `\t`.