refact(youtube): Decouple channel fetching and URL parsing logic (#1478)
* feat(youtube): Add ytChannelID interface

This interface has been added, with the aim of splitting the fetching
logic currently handled by `Plugin.getYtChannel`, into methods specific
to the type of ID parsed, and separate from their parsing logic, as well
as eliminating the `ytUrlType` type, and associated constants,
in favour of using it as the return type for `Plugin.parseYtUrl`.

* feat(youtube): Add videoID newtype

`videoID` is a newtype, that encapsulates strings representing
YouTube video ID, and features an implementation of the
`getChannelList` method, which fetching data about the video
it refers to, using it to add the channel ID to a `youtube.ChannelsListCall`.

This commit also adds the `listParts` var, as this will be a common
value, also used by other fetching methods.

* feat(youtube): Add channelID newtype

`channelID` is a newtype, that encapsulates strings representing
a YouTube channel ID, and features an implementation of the
`getChannelList` method, adds the ID to a `youtube.ChannelsListCall`.

* feat(youtube): Add userID newtype

`userID` is a newtype, that encapsulates strings which represent
a YouTube username, and features an implementation of the
`getChannelList` method, which adds the username as the `forUsername`
argument, of an `youtube.ChannelsListCall`.

* feat(youtube): Add searchChannelID newtype

`searchChannelID` is a newtype, that encapsulates strings which represent
a YouTube handle, or custom URL, and features an implementation of the
`getChannelList` method, which uses `youtube.SearchServiceCall` to try
and look up channels with the specified string, and add their ID to a
`youtube.ChannelListCall`.

* feat(youtube): Return ytChannelID from parseYtUrl

This commit updates `Plugin.parseYtUrl` to return the `ytChannelID`
interface, rather than `ytUrlType` and an ID string. This change has
been made, to support decoupling `Plugin.parseYtUrl` from
`Plugin.getYtChannel` in future, as well as supporting improved error
reporting.

* feat(youtube): Return *videoID from parseYtVideoID

Update `Plugin.parseYtVideoID` to return `*videoID`, a type implementing
the `ytChannelID` interface, inline with changes to `Plugin.parseYtUrl`.

* refact(youtube): Move response logic to HandleNew

This commit updates `Plugin.HandleNew` to use the redesigned
`Plugin.parseYtUrl` and the new `ytChannelID` interface,
as well as transplanting some of the response handling logic from
`Plugin.getYtChannel`, into `Plugin.HandleNew`, to accommodate these
changes.

* refact(youtube): Remove ytUrlType and constants

Removes the `ytUrlType` type and associated constants, due to no longer
being used.

* refact(youtube): Remove getYtChannel

Removes the `Plugin.getYtChannel` method, due to no longer being used.

* refact(youtube): Move HandleNew domain validation

Moves the domain validation performed in `Plugin.HandleNew`, to
`Plugin.parseYtUrl`. This avoids duplicating the validation logic for
`youtu.be`, which is already required for parsing one of the variants of
`Plugin.parseYtUrl`, as well as avoding flattening it with errors
produced by `Url.Parse`.

* refact(youtube): Split segments from path

Swaps `strings.Split(channelUrl.Path, "/")` for `strings.Split(path, "/")`,
due to the leading `/` being trimmed from `path` at the beginning of
`Plugin.parseYtUrl`, and avoiding an empty first item in `pathSegments`.

* feat(youtube): Wrap formatted strings in ""

Updates `fmt.Errorf` format strings, to wrap injected strings in `""`,
to ensure included strings are visible, if a string is empty.

* fix(youtube): Ensure - is escaped in RegEx

This commit makes sure hyphens are escaped in `ytVideoIDRegex` and
`ytChannelIDRegex`, as these are considered metacharacters in character
groups.

* refact(youtube): Switch to handle vaidation

Due to the implementation of `searchChannelID.getChannelList` escaping
the contents of queries, it is no longer necessary to strip the leading
`@` when parsing handle URLs, so this updates the RegEx and parsing
logic accordingly.

* refact(youtube): Switch to ID validation

Due to the way `ytVideoIDRegex` and `ytChannelIDRegex` are designed,
when using `Regex.FindString`, the contents of the string produced,
shouldn't differ from the contents of the string given as an argument,
meaning `Regex.MatchString` can be used instead, simplifying the logic.

* refact(youtube): Remove semi-redundant nil checks

This commit removes `err == nil` checks in the `getChannelList` method
implementations in `youtube/feed.go`, to improve the functions
readability, as doing so shouldn't result in erroneous behaviour.

Consequently, this will result in the caller name fetching behaviour in
`common.ErrWithCaller` being invoked unnecessarily, but judging by
design patterns in error handling functions, this should be addressed
by introducing a `nil` check to `common.ErrWithCaller`.

* fix(youtube): Remove unused strings import

* refact(youtube): Use idiomatic reciever names

Updates the names of the reciever variables in `videoID.getChannelList`,
`channelID.getChannelList`, `userID.getChannelList` and
`searchChannelID.getChannelList`, to use the first letter of the type
names, inline with Go idioms.

* refact(youtube): Make ytChannelID newtypes strings

This commit makes the newtypes introduced by this branch, `string`s,
rather than `struct`s, alongside updating any usage of the types
accordingly, including updating the return type of
`Plugin.parseVideoID`.

* refact(youtube): Swap \"%s\" for %q

In commit `859d1e6`, I wrapped `%s` flags in error formatting messages,
in `""`, to ensure empty strings would still be visible in error
messages.

After some discussion though, it was decided that using the `%q` flag
instead, was preferable, as this would achieve a similar effect, but
would ensure any characters which may result in confusing formatting
when printing the error, such as newlines and tabs, would instead
be replaced by their escape sequence equivilents, such as `\n` and `\t`.