feat(youtube): Add playlist URL support (#1488)
* feat(youtube): Add playlistID type

This commit adds the `playlistID` newtype, and accompanying
`playlistID.getChannelList` method, for representing a YouTube playlist
ID, and fetching the channel which created it.

* feat(youtube): Add playlist URL parsing logic

This commit adds support for parsing YouTube playlist URLs,
to `Plugin.parseYtUrl`. This includes adding a new Regular Expression
for validating the ID format, and updating the formatting for
surrounding expressions.