moderation: add mod convenience functions for CCs
Add moderation convenience functions to the custom command template
functions.

Currently, users have to use `exec` and `execAdmin` to execute such
tasks. This and future commits aim to get rid of this status quo and
instead provide functions to handle this, freeing up the exec calls for
other, rarely used commands.

This was a a suggestion on the support server, found here:
https://canary.discord.com/channels/166207328570441728/356486960417734666/1108833432953438319

Signed-off-by: Luca Zeuch <l-zeuch@email.de>