moderation: add cc convenience functions (#1516)
* common/templates: export TargetUserID

Export TargetUserID (former targetUserID) because we're going to use it
later outside of its package. There is zero reason to have duplicated
code in this instance.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* moderation: add mod convenience functions for CCs

Add moderation convenience functions to the custom command template
functions.

Currently, users have to use `exec` and `execAdmin` to execute such
tasks. This and future commits aim to get rid of this status quo and
instead provide functions to handle this, freeing up the exec calls for
other, rarely used commands.

This was a a suggestion on the support server, found here:
https://canary.discord.com/channels/166207328570441728/356486960417734666/1108833432953438319

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* moderation: rename tmplextension to tmplextensions

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* moderation/tmplextensions: add various checks

Add various checks to the `tmplGetWarns` function such as a call counter
specific to this and future functions, a test for sane input to
templates.TargetUserID as well as avoiding to list expired logs in the
result.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* moderation/tmplextensions: be better at erroring

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>