lib/dstate: add GuildSet.IconURL method
Add a IconURL method to the dstate.GuildSet struct to retrieve the
guild's icon URL.

Currently, we have to use the following code to retrieve a guild's icon
in custom commands:

```
{{$ex := or (and (hasPrefix "a_" .Guild.Icon) "gif" ) "png" }}
{{$icon := printf "https://cdn.discordapp.com/icons/%d/%s.%s?size=1024" .Guild.ID .Guild.Icon $ex }}
```

With this change, users will be able to condense the above down to the
following:

```
{{ $icon := .Guild.IconURL "1024" }}
```
Which quite obviously is a lot cleaner and space-efficient.

I wasn't able to find a relevant suggestion to back myself up with, but
I believe this addition was rather long due and will make it a bit
easier for our users in the long run.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>