frontend: link to Discord statuspage for incidents (#1544)
* frontend: link to Discord statuspage for incidents

Link to the relevant incident page when we are forwaring an incident
from Discord on the control panel, so there isn't much left to ask.

Relevant suggestion on the support server:
https://discord.com/channels/166207328570441728/356486960417734666/909083451704115231

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* frontend/cp_main: use string formatting instead

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* frontend/cp_main: camelCase variable naming

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* frontend: add new addAlertHTML func

Add a new addAlertHTML function to use arbitrary HTML when adding
alerts. This is necessary because the addAlert function uses jQuery's
$.text() function, which escapes HTML.

This new function uses $.html() instead. For security and compatibility,
we keep the old function and only use addAlertHTML when absolutely
necessary.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* fixup! frontend: add new addAlertHTML func

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>