feat(logs): Add login for members only logs error
Users have continually expressed confusion regarding the error used
when attempting to open a message log for a server, where logs
are restricted to members only. The current message implies the user
isn't a member of the server, even if they are.

This commit attempts to rectify this, by adding a prompt
suggesting the user should use a login link, if they are a member.