feat(setstatus): Implement custom status in discordgo and in setstatus
Now that custom status is available to bots, implement it in discordgo using the GameType struct, and adding Competing while we're at it.
Additionally, add flags to setstatus to permit the command user to set the online status of YAGPDB, as well as choose the type of presence.
Also, rename GameType and associated values to ActivityType to more closely match the API.

Signed-off-by: Clari <clarinet.puppy@gmail.com>
Tested-by: Luca Zeuch <l-zeuch@email.de>