customcommands/web: add database view page
Add a database view page to the control panel, allowing users to view
their custom command database, and, as necessary, search for entries
matching specific criteria.

The output is, no matter the request, paginated to max 100 entries per
page, which is in-line with the maximum one can achieve via a `dbGetPattern`
call in custom commands.

I've also decided to add delete functionality for individual entries,
such that if a stray entry that is no longer needed is found, it can be
deleted without resorting to a custom command / `evalcc`.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>