discordgo: fix button message component emoji errors
Change discordgo.Button.Emoji to *discordgo.ComponentEmoji and include
the omitempty tag.

Discord no longer allows the `emojis` field in button objects to be
null, and instead require it be absent. This commit adds the "omitempty"
tag to the Emoji field, and changes the value to a pointer.

See also the suggested fix from the Discord Gophers server:
https://discord.com/channels/118456055842734083/155361364909621248/1192299819012722748