cc: allow duplicate keys in complexMessage<Edit>
allow multiple of the same key to be passed to complexMessage so that
embeds and components may be declared multiple times, each case
appending the input to the message.
illustrated here: https://github.com/botlabs-gg/yagpdb/assets/110698921/c2e1e9fa-8f36-4ddd-8fd0-d45f4ad945d6
this is mainly implemented for buttons
and menus since embed already uses a
slice of embeds and this can be done
even better using that method. this
is not as simple with buttons and menus
because we need to be able to customize
the layout of the component rows.
only keeping it the same for embed
field for consistency.

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>