cc/interactions: allow file in response + followup
add the ability for a file specified in complexMessage to be sent in an
interaction response or followup. Technically a fix because I thought it
would already work but never tested it. Feauring a slightly altered
upstream pull of discordgo for (s *Session) CreateInteractionResponse.
It still diverges from upstream slightly, however only enough to avoid
rebasing the entirety of restapi.go.

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>