customcommands: compare underlying types for number conversion
Use reflection to compare the underlying type of arguments to convert
numbers more reliably.
Most users would expect `toInt`, `toInt64`, and `toFloat` to convert any
number, whether or not it's a defined type.
This is currently worked around by adding explicit cases for some types,
however some possible cases are missing and this approach is future-proof.

This addresses suggestion #2142 on the support server:
https://canary.discord.com/channels/166207328570441728/356486960417734666/1158622033001844746

Signed-off-by: Galen CC <galen8183@gmail.com>