fix(discordgo): Use headers for kick reason
In previous versions, the Discord API accepted an audit log reason
when endpoints supported them, by checking for a `reason` key
appended to the request URL. Discord have since updated their policies,
to use the `X-Audit-Log-Reason` header for this purpose.

`Session.GuildMemberDeleteWithReason` is still based on
this old request structure, resulting in audit log reasons
failing to be attached when kick requests were made.
This commit updates the function to use the new request structure.