configstore: simplify code
Our goal is to remove the configstore package and inline uses at callsites (which will then allow migration to sqlboiler.) This commit removes unused parts of configstore to make it easier to see its function.

Stripping away all the abstraction layers, then, we see that configstore -- as used in the codebase today -- is just a ccache over Postgres, where queries to the database layer are retried up to 1000 times.

The next commit will remove configstore and implement the cache+Postgres queries directly in the two clients of configstore -- the notifications and moderation plugins.