all: futureproof viewperms and template Permissions map by autogenerating perm name to bit map (#1649)
* lib/discordgo: remove deprecated permission names

* lib/discordgo: add AllPermissions and generate PermissionString function

* common: implement util.HumanizePermissions using discordgo.PermissionName

* templates: generate context Permissions map using discordgo.PermissionName

As a happy side effect, this new implementation means the Permissions map contains a number of new permissions that were previously inaccessible in templates, such as but not limited to MODERATE_MEMBERS.

* dstate: remove duplicate ManageGuild in ChannelPermsMask

* discordgo: switch Permission constants to use left-shift expressions

* discordgo: add new permissions

This commit is a nice example of the benefits of the refactor. After changing `permission.go`, running `go generate` will update `permission_name.go`. In turn `viewperms` and the `Permissions` context map, will support the added permissions, without extra work.